+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerField.test.jsx
@@ -1,26 +1,45 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import DatePickerField from '../../../DatePickerField';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class DatePickerExample extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { isInvalid: false };
-    this.handleButtonClick = this.handleButtonClick.bind(this);
+    this.state = {
+      isInvalid: false,
+      isIncomplete: false,
+      required: false,
+    };
+    this.handleInvalidButtonClick = this.handleInvalidButtonClick.bind(this);
+    this.handleIncompleteButtonClick = this.handleIncompleteButtonClick.bind(this);
   }
 
-  handleButtonClick() {
+  handleInvalidButtonClick() {
     this.setState(prevState => ({
       isInvalid: !prevState.isInvalid,
     }));
   }
 
+  handleIncompleteButtonClick() {
+    this.setState(prevState => ({
+      isIncomplete: !prevState.isIncomplete,
+      required: !prevState.required,
+    }));
+  }
+
   render() {
     return (
-      <div>
-        <button type="button" id="validity-toggle" onClick={this.handleButtonClick}>Toggle Validity</button>
+      <div className={cx('content-wrapper')}>
+        <button type="button" id="validity-toggle" onClick={this.handleInvalidButtonClick}>Toggle Validity</button>
+        <button type="button" id="incomplete-toggle" onClick={this.handleIncompleteButtonClick}>Toggle Incomplete</button>
         <DatePickerField
           label="Enter Date"
           isInvalid={this.state.isInvalid}
+          isIncomplete={this.state.isIncomplete}
+          required={this.state.required}
           error="Error message."
           name="date-input"
           datePickerId="default"