+++ packages/terra-popup/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Updated docs.
 
+6.15.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 6.14.0 - (October 3, 2019)
 ------------------
 ### Changed